+++ R/addEta.R
@@ -25,7 +25,7 @@
 #' @param priorName logical, if TRUE, the parameter name specified in
 #'   `eta` will be used to add the eta value prior name is used
 #'   instead of the left handed side of the equation.
-#' @param etaCombineType the option for the how to ccombine the eta
+#' @param etaCombineType the option for the how to combine the eta
 #'   with the parameter name.  Can be: "default", "snake", "camel",
 #'   "dot", "blank"
 #' @return The model with eta added to the requested parameters