+++ inst/modeldb/therapeuticArea/oncology/tgi_sat_logistic.R
@@ -1,0 +1,34 @@
+tgi_sat_logistic <- function() {
+  description <- "One compartment TGI model with with exponential tumor growth that decelerates linearly, with saturation."
+  ini({
+    lts0 <- 0.3; label("Initial tumor size (TS0)") 
+    ltsmax <- 0.9; label("Maximum tumor size at saturation (TSmax)")
+    lka <- 0.45 ; label("Absorption rate (Ka)")
+    lcl <- 1 ; label("Clearance (CL)")
+    lvc  <- 3.45 ; label("Central volume of distribution (V)")
+    lkgl <- 0.7; label("Zero-order linear growth rate")
+    CcpropSd <- 0.5 ; label("PK proportional residual error (fraction)")
+    tumorSizepropSd <- 0.5 ; label("Tumor size proportional residual error (fraction)")
+    tumorSizeaddSd <- 30 ; label("Tumor size additive residual error (tumor volume)")
+  })
+  model({
+    ts0 <- exp(lts0)
+    tsmax <- exp(ltsmax)
+    ka <- exp(lka)
+    cl <- exp(lcl)
+    vc  <- exp(lvc)
+    kgl <- exp(lkgl)
+    
+    kel <- cl / vc
+    tumorSize(0) <- ts0
+    
+    
+    d/dt(depot) <- -ka*depot
+    d/dt(central) <- ka*depot-kel*central
+    d/dt(tumorSize) <- kge*tumorSize*(1-(tumorSize/tsmax))
+    
+    Cc <- central / vc
+    Cc ~ prop(CcpropSd)
+    tumorSize ~ prop(tumorSizepropSd) + add(tumorSizeaddSd)
+  })
+}